% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismember.R
\name{dismember}
\alias{dismember}
\title{Dismember data element into its distinct components}
\usage{
dismember(df)
}
\arguments{
\item{df}{dataframe to adjust}
}
\description{
Dismember data element into its distinct components
}
