% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_analytics.R
\name{pull_analytics}
\alias{pull_analytics}
\title{Run an Analytics API call on DATIM}
\usage{
pull_analytics(url, baseurl = "https://pisa.datim.org/", extract = "rows")
}
\arguments{
\item{url}{api url, after base url}

\item{baseurl}{default = "https://pisa.datim.org/"}

\item{extract}{what element of the json file do you want to use, rows, dataElement}
}
\description{
Run an Analytics API call on DATIM
}
