% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_ind.R
\name{pull_ind}
\alias{pull_ind}
\title{Pull indicator data from DATIM}
\usage{
pull_ind(ind, pd = "2018Q1")
}
\arguments{
\item{ind}{indicator to pull down}

\item{pd}{period to pull, eg "2018Q1" or multiple "2017Q4;2018Q1"}
}
\description{
Pull indicator data from DATIM
}
